[publish][m #32]: Added more Readme Test
Added Test case for Readme warning, Extension
